chore(warnings): unify startup info init on Windows and adjust unsigned literal checks
The Windows process_launcher initialization now uses a zero-initialized
STARTUPINFO followed by an explicit cb assignment, replacing the previous
STARTUPINFO si{ sizeof(si) }. This modification is intended only to silence
initialization warnings emitted by MinGW-w64 GCC and does not change runtime
behavior.

Several test comparisons in test_pod_sys.cpp such as REQUIRE(h1 != -1) were
updated to REQUIRE(h1 != static_cast<std::uint64_t>(-1)) to address GCC 15+
diagnostics regarding implicit conversion from a signed literal to an unsigned
type. These adjustments are strictly warning-suppression changes with no
functional impact.

CI targets (triggered on push):

* Linux (Ubuntu / GCC)
* Darwin (macOS / LLVM 20)
* Windows (MSYS2 UCRT64 / MinGW-w64)

+++ tests/test_pod_sys.cpp
@@ -838,13 +838,13 @@ TEST_CASE("string_view hash reflects exact character content") {
         auto h5 = sv1.hash(c_hash::murmur64);
         auto h6 = sv1.hash(c_hash::xxhash64);
 
-        REQUIRE(h1 != -1);
-        REQUIRE(h2 != -1);
-        REQUIRE(h3 != -1);
-        REQUIRE(h4 != -1);
-        REQUIRE(h5 != -1);
-        REQUIRE(h5 != -1);
-        REQUIRE(h6 != -1);
+        REQUIRE(h1 != static_cast<std::uint64_t>(-1));
+        REQUIRE(h2 != static_cast<std::uint64_t>(-1));
+        REQUIRE(h3 != static_cast<std::uint64_t>(-1));
+        REQUIRE(h4 != static_cast<std::uint64_t>(-1));
+        REQUIRE(h5 != static_cast<std::uint64_t>(-1));
+        REQUIRE(h5 != static_cast<std::uint64_t>(-1));
+        REQUIRE(h6 != static_cast<std::uint64_t>(-1));
 
         // Same view, multiple algorithms must differ
         REQUIRE(h1 != h2);
